[Chore] #134 - DateBoard/Detail/Views SettingView - View 이름 변경, 코드 중복 최소화, Mark  추가
